1-24-25

+++ database/migrations/2024_11_08_030836_create_postpartum_table.php
@@ -11,40 +11,45 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('postpartum', function (Blueprint $table) {
-            $table->bigIncrements('postpartumId');  // Custom primary key
-            $table->unsignedBigInteger('prenatalConsultationDetailsID');
-            $table->unsignedBigInteger('id');
-            $table->string('lastName', 100);
-            $table->string('firstName', 100);
-            $table->string('middleName', 100);
-            $table->string('sex', 10);
-            $table->string('prenatalDelivered', 100);
-            $table->string('placeDelivered', 100);
-            $table->string('modeOfDelivery', 100);
-            $table->decimal('birthLength', 5, 2);
-            $table->decimal('birthWeight', 5, 2);
-            $table->date('deliveryDate');
-            $table->time('deliveryTime');
-            $table->string('attendantBirth', 100);
-            $table->date('dateInitiatedBreastfeeding');
-            $table->time('timeInitiatedBreastfeeding');
-            $table->date('dateOfPostpartumVisitTwentyFourHoursDelivery');
-            $table->date('dateOfPostpartumVisitOneWeekDelivery');
-            $table->date('dateVitaminA');
-            $table->date('dateIronGiven');
-            $table->decimal('noIronGiven', 5, 2);
-            $table->string('dangerSignsMother', 100);
-            $table->string('dangerSignsBaby', 100);
-            $table->timestamps();
-
-            $table->foreign('prenatalConsultationDetailsID')->references('prenatalConsultationDetailsID')->on('prenatal_consultation_details')->onDelete('cascade');
-
-            $table->foreign('id')
-            ->references('id')
-            ->on('users')
-            ->onDelete('cascade');
-        });
+        if (!Schema::hasTable('postpartum')) {
+            Schema::create('postpartum', function (Blueprint $table) {
+                $table->bigIncrements('postpartumId');  // Custom primary key
+                $table->unsignedBigInteger('prenatalConsultationDetailsID');
+                $table->unsignedBigInteger('id');
+                $table->string('lastName', 100);
+                $table->string('firstName', 100);
+                $table->string('middleName', 100);
+                $table->string('sex', 10);
+                $table->string('prenatalDelivered', 100);
+                $table->string('placeDelivered', 100);
+                $table->string('modeOfDelivery', 100);
+                $table->decimal('birthLength', 5, 2);
+                $table->decimal('birthWeight', 5, 2);
+                $table->date('deliveryDate');
+                $table->time('deliveryTime');
+                $table->string('attendantBirth', 100);
+                $table->date('dateInitiatedBreastfeeding');
+                $table->time('timeInitiatedBreastfeeding');
+                $table->date('dateOfPostpartumVisitTwentyFourHoursDelivery');
+                $table->date('dateOfPostpartumVisitOneWeekDelivery');
+                $table->date('dateVitaminA');
+                $table->date('dateIronGiven');
+                $table->decimal('noIronGiven', 5, 2);
+                $table->string('dangerSignsMother', 100);
+                $table->string('dangerSignsBaby', 100);
+                $table->timestamps();
+    
+                $table->foreign('prenatalConsultationDetailsID')
+                    ->references('prenatalConsultationDetailsID')
+                    ->on('prenatal_consultation_details')
+                    ->onDelete('cascade');
+    
+                $table->foreign('id')
+                    ->references('id')
+                    ->on('users')
+                    ->onDelete('cascade');
+            });
+        }
     }
 
     /**
